+++ src/main/java/ch/jtaf/ui/view/OrganizationsView.java
@@ -1,15 +1,15 @@
 package ch.jtaf.ui.view;
 
 import ch.jtaf.db.tables.records.OrganizationRecord;
-import ch.jtaf.ui.security.OrganizationHolder;
 import ch.jtaf.ui.dialog.OrganizationDialog;
 import ch.jtaf.ui.layout.MainLayout;
+import ch.jtaf.ui.security.OrganizationHolder;
 import com.vaadin.flow.component.UI;
 import com.vaadin.flow.component.button.Button;
 import com.vaadin.flow.component.button.ButtonVariant;
+import com.vaadin.flow.component.confirmdialog.ConfirmDialog;
 import com.vaadin.flow.component.grid.ColumnTextAlign;
 import com.vaadin.flow.component.grid.Grid;
-import com.vaadin.flow.component.html.H1;
 import com.vaadin.flow.component.notification.Notification;
 import com.vaadin.flow.component.orderedlayout.HorizontalLayout;
 import com.vaadin.flow.component.orderedlayout.VerticalLayout;
@@ -19,13 +19,15 @@ import org.jooq.DSLContext;
 import org.jooq.exception.DataAccessException;
 import org.springframework.security.core.context.SecurityContextHolder;
 
+import java.io.Serial;
+
 import static ch.jtaf.db.tables.Organization.ORGANIZATION;
 import static ch.jtaf.db.tables.OrganizationUser.ORGANIZATION_USER;
-import static ch.jtaf.db.tables.SecurityUser.SECURITY_USER;
 
 @Route(layout = MainLayout.class)
 public class OrganizationsView extends VerticalLayout implements HasDynamicTitle {
 
+    @Serial
     private static final long serialVersionUID = 1L;
 
     private final transient DSLContext dsl;
@@ -36,14 +38,13 @@ public class OrganizationsView extends VerticalLayout implements HasDynamicTitle
 
         setHeightFull();
 
-        add(new H1(getTranslation("My.Organizations")));
-
         OrganizationDialog dialog = new OrganizationDialog(getTranslation("Organization"));
 
         Button add = new Button(getTranslation("Add"));
         add.addClickListener(event -> dialog.open(ORGANIZATION.newRecord(), this::loadData));
 
         grid = new Grid<>();
+        grid.getClassNames().add("rounded-corners");
         grid.setHeightFull();
 
         grid.addColumn(OrganizationRecord::getOrganizationKey).setHeader(getTranslation("Key")).setSortable(true);
@@ -51,20 +52,29 @@ public class OrganizationsView extends VerticalLayout implements HasDynamicTitle
 
         grid.addComponentColumn(organizationRecord -> {
             Button select = new Button(getTranslation("Select"));
+            select.addThemeVariants(ButtonVariant.LUMO_PRIMARY);
             select.addClickListener(event -> {
                 OrganizationHolder.setOrganization(organizationRecord);
                 UI.getCurrent().navigate(SeriesListView.class);
             });
 
             Button delete = new Button(getTranslation("Delete"));
-            delete.addThemeVariants(ButtonVariant.LUMO_ERROR);
+            delete.addThemeVariants(ButtonVariant.LUMO_PRIMARY, ButtonVariant.LUMO_ERROR);
             delete.addClickListener(event -> {
-                try {
-                    dsl.attach(organizationRecord);
-                    organizationRecord.delete();
-                } catch (DataAccessException e) {
-                    Notification.show(e.getMessage());
-                }
+                ConfirmDialog confirmDialog = new ConfirmDialog(getTranslation("Confirm"),
+                    getTranslation("Are.you.sure"),
+                    getTranslation("Delete"), e -> {
+                    try {
+                        dsl.attach(organizationRecord);
+                        organizationRecord.delete();
+                    } catch (DataAccessException ex) {
+                        Notification.show(ex.getMessage());
+                    }
+                },
+                    getTranslation("Cancel"), e -> {
+                });
+                confirmDialog.setConfirmButtonTheme("error primary");
+                confirmDialog.open();
             });
 
             HorizontalLayout horizontalLayout = new HorizontalLayout(select, delete);
@@ -82,11 +92,9 @@ public class OrganizationsView extends VerticalLayout implements HasDynamicTitle
 
     private void loadData() {
         var organizations = dsl
-            .select()
-            .from(ORGANIZATION)
-            .join(ORGANIZATION_USER).on(ORGANIZATION_USER.ORGANIZATION_ID.eq(ORGANIZATION.ID))
-            .join(SECURITY_USER).on(SECURITY_USER.ID.eq(ORGANIZATION_USER.USER_ID))
-            .where(SECURITY_USER.EMAIL.eq(SecurityContextHolder.getContext().getAuthentication().getName()))
+            .select(ORGANIZATION_USER.organization().fields())
+            .from(ORGANIZATION_USER)
+            .where(ORGANIZATION_USER.securityUser().EMAIL.eq(SecurityContextHolder.getContext().getAuthentication().getName()))
             .fetch().into(ORGANIZATION);
 
         grid.setItems(organizations);
@@ -94,6 +102,6 @@ public class OrganizationsView extends VerticalLayout implements HasDynamicTitle
 
     @Override
     public String getPageTitle() {
-        return "JTAF - " + getTranslation("Organizations");
+        return getTranslation("Organizations");
     }
 }
